+++ changelog.md
@@ -1,5 +1,6 @@
 # v1.1.1
 - Added toggle setting `Geode Mod Pop-up`
+- Fixed Geode loader mod not appearing in favorites list if added to favorites
 - Tweaked favorites UI on Geode mod pop-up
 
 # v1.1.0